+++ croud/__main__.py
@@ -23,7 +23,12 @@ import sys
 
 import colorama
 
-from croud.clusters.commands import clusters_deploy, clusters_list, clusters_scale
+from croud.clusters.commands import (
+    clusters_delete,
+    clusters_deploy,
+    clusters_list,
+    clusters_scale,
+)
 from croud.cmd import (
     cluster_id_arg,
     cluster_name_arg,
@@ -294,7 +299,18 @@ command_tree = {
                     ),
                 ],
                 "resolver": clusters_scale,
-            }
+            },
+
+            "delete": {
+                "help": "Delete the specified cluster.",
+                "extra_args": [
+                    lambda req_opt_group, opt_opt_group: cluster_id_arg(
+                        req_opt_group, opt_opt_group, True
+                    ),
+                    yes_arg
+                ],
+                "resolver": clusters_delete,
+            },
         },
     },
     "products": {